# matrix-service/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "Cyberpunk 2077 hacking mini game",
        "contact": {}
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/GetSequence": {
            "get": {
                "description": "Get new matrix and sequences for hacking",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Main tools"
                ],
                "summary": "GetSequence",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "id of Matrix",
                        "name": "id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
        "/GetSequenceSugar": {
            "get": {
                "description": "Get new matrix and sequences for hacking with human representation",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Additional methods"
                ],
                "summary": "GetSequence For Human",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "id of Matrix",
                        "name": "id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
        "/Hack": {
            "post": {
                "description": "Try to hack matrix by your own",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Main tools"
                ],
                "summary": "Hack",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "id of matrix",
                        "name": "matrix_id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "path with clear coordinates to hack matrix",
                        "name": "path",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {}
            }
        }
    }
}